python/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai-whisper>=20231117
python-multipart>=0.0.6
openai>=1.3.0
python-dotenv>=1.0.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0

# OpenAI Integration
openai>=1.3.0
python-dotenv>=1.0.0
openai>=1.3.0
